ocr/backend/requirements.txt
easyocr
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
Pillow
